Fix computation of cursor positions.
vim.current.window.cursor returns (row, col) where row is the line number and
col is the byte offset on the current line. This works fine for one-byte
characters since col moves one byte with every character which is same as
cursor behaves. However, for multi-byte characters, col moves more than one
byte for a character. So, the cursor position is not compatible with the col.